# services/main-backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/allergies": {
      "get": {
        "operationId": "AllergyController_getAllergy",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Trả về danh sách allergy"
          },
          "401": {
            "description": "Token không hợp lệ hoặc đã hết hạn"
          },
          "404": {
            "description": "User not found."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Lấy danh sách allergy của user hiện tại",
        "tags": [
          "Allergy"
        ]
      },
      "patch": {
        "operationId": "AllergyController_updateAllergy",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "ingredientIds"
                ],
                "properties": {
                  "ingredientIds": {
                    "type": "array",
                    "items": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "example": [
                      1,
                      2,
                      3
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cập nhật allergy thành công"
          },
          "401": {
            "description": "Token không hợp lệ hoặc đã hết hạn"
          },
          "404": {
            "description": "User or ingredient not found."
          },
          "409": {
            "description": "Ingredient conflict with favorite list."
          },
          "422": {
            "description": "Invalid request payload."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Cập nhật danh sách allergy của user hiện tại",
        "tags": [
          "Allergy"
        ]
      }
    },
    "/api/v1/favorite-ingredients": {
      "get": {
        "operationId": "FavoriteIngredientController_getFavoriteIngredient",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Trả về danh sách favorite ingredient"
          },
          "401": {
            "description": "Token không hợp lệ hoặc đã hết hạn"
          },
          "404": {
            "description": "User not found."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Lấy danh sách favorite ingredient của user hiện tại",
        "tags": [
          "Favorite Ingredient"
        ]
      },
      "patch": {
        "operationId": "FavoriteIngredientController_updateFavoriteIngredient",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "ingredientIds"
                ],
                "properties": {
                  "ingredientIds": {
                    "type": "array",
                    "items": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "example": [
                      1,
                      2,
                      3
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cập nhật favorite ingredient thành công"
          },
          "401": {
            "description": "Token không hợp lệ hoặc đã hết hạn"
          },
          "404": {
            "description": "User or ingredient not found."
          },
          "409": {
            "description": "Ingredient conflict with allergy list."
          },
          "422": {
            "description": "Invalid request payload."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Cập nhật danh sách favorite ingredient của user hiện tại",
        "tags": [
          "Favorite Ingredient"
        ]
      }
    },
    "/api/v1/options/diet-types": {
      "get": {
        "operationId": "OptionsController_getDietTypes",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Trả về danh sách diet types"
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "summary": "Lấy danh sách diet types",
        "tags": [
          "Options"
        ]
      }
    },
    "/api/v1/options/diet-types/{id}": {
      "get": {
        "operationId": "OptionsController_getDietTypeById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Trả về diet type theo id của diet type"
          },
          "404": {
            "description": "Diet type not found."
          },
          "422": {
            "description": "Invalid option id."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "summary": "Lấy chi tiết diet type theo id của diet type",
        "tags": [
          "Options"
        ]
      }
    },
    "/api/v1/options/goals": {
      "get": {
        "operationId": "OptionsController_getGoals",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Trả về danh sách goals"
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "summary": "Lấy danh sách goals",
        "tags": [
          "Options"
        ]
      }
    },
    "/api/v1/options/goals/{id}": {
      "get": {
        "operationId": "OptionsController_getGoalById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Trả về goal theo id của goal"
          },
          "404": {
            "description": "Goal not found."
          },
          "422": {
            "description": "Invalid option id."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "summary": "Lấy chi tiết goal theo id của goal",
        "tags": [
          "Options"
        ]
      }
    },
    "/api/v1/options/cuisine-types": {
      "get": {
        "operationId": "OptionsController_getCuisineTypes",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Trả về danh sách cuisine types"
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "summary": "Lấy danh sách cuisine types",
        "tags": [
          "Options"
        ]
      }
    },
    "/api/v1/options/cuisine-types/{id}": {
      "get": {
        "operationId": "OptionsController_getCuisineTypeById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Trả về cuisine type theo id của cuisine type"
          },
          "404": {
            "description": "Cuisine type not found."
          },
          "422": {
            "description": "Invalid option id."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "summary": "Lấy chi tiết cuisine type theo id của cuisine type",
        "tags": [
          "Options"
        ]
      }
    },
...
  },
  "info": {
    "title": "KitchenMind API",
    "description": "Tài liệu hướng dẫn và kiểm thử API cho hệ thống Meal Planner\n\n## Xác thực\nCác endpoint được bảo vệ yêu cầu JWT Bearer token.\nĐăng nhập qua Google (`GET /api/auth/google`) để nhận `accessToken`,\nsau đó nhấn **Authorize** và nhập token vào ô `Bearer <token>`.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Nginx Gateway Local"
    },
    {
      "url": "http://localhost:3000",
      "description": "Direct Main-Backend Port"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Nhập JWT token (không cần tiền tố Bearer)",
        "in": "header"
      }
    },
...
  }
}